backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pdfplumber==0.11.4
spacy==3.7.2
scikit-learn==1.3.2
requests==2.31.0